moved class StatusOption
